Add serialization/deserialization for config & decisions
Since we're gonna need to parse the config settings and the decisions,
we should bite the bullet and do full-on serialization/deserialization
for them. This will make evaluating/reproducing results much easier.
As a sideffect, we no longer have to manually parse the config, as it's
in JSON format, and we simply deserialize it.