Make skipping traces apply to loading failures too
When we decide to skip a trace, we shouldn't just do so when the file
isn't there; sometimes the file is *partially* there, and we will fail
trying to read it all; in that case, we can just skip it and wait for
the next run to read it.

This was seen mostly when testing `yes`, maybe it has to do something
with the fact that the fuzzer is constantly terminating the target
program.