Time each decision made by the oracle
In order to get stats about discovery time (such as "time to first
backdoor"), it would be useful to attach a timestamp to each decision
made by the oracle. That way, we can figure out how much time has
elapsed between the start of the detection campaign and a given
decision.