Allow for multiple fuzzers
In both the seed and the run phase, we should be able to perform
parallel fuzzing, to be able to detect backdoors in larger programs. To
do this, instead of having a single fuzzer instance in both phases, we
can now have a vector of fuzzer instances in each phase. That way, the
user can specify as many fuzzers as they wish in the config for each
phase.