Respawn dead run-phase fuzzers
Sometimes, we'll start the run phase with some fuzzers not having
spawned correctly. This *sometimes* happens because of memory issues; in
that case, we should try to respawn them, giving them enough time to
come back.

The stats (how many fuzzers are alive) should also be displayed on the
TUI.