Refactor trace UIDs
ROSA used to be based on the file names generated by the fuzzer for the
different inputs and accompanying trace dumps. This worked okay with the
SIMPLE_FILES setting on the AFL++ side, guaranteeing that seeds will not
be seen as new traces, and that truly new traces will not be ignored.

This, obviously, was very flaky; even worse, it turns out that the
SIMPLE_FILES setting does not play well with parallel fuzzing, which is
essential for our benchmarks. In order to kill two birds with one stone,
inputs (and thus traces) are now named after the hash of their edges and
syscalls. Obviously collisions are possible, but they are worth it to be
able to know which traces have been seen before regardless of what the
fuzzer named their files.

As a nice side-effect, we now only load traces we care about: any
duplicates will be eliminated by having the same hash as other known
traces, which should reduce the number of false positives.