trace.rs: filter out duplicate traces
Due to a (currently unknown) bug in the trace dump part of AFL++, some
traces appear to have different hashes even though they have the exact
same test inputs. Some of this can be filtered out quickly (done in this
commit), but this is probably happening even when the test inputs are
not exactly the same.