rosa.rs: add multi-fuzzer trace harvesting back
Commit 25522964 removed the feature of collecting traces from every
fuzzer (in both the seed and the detection phase), because we expect all
traces to eventually be shared with the main fuzzer if they are
interesting. However, in reality that might take longer than expected;
in the interest of performing inference as soon as possible, we'll query
all fuzzers for traces. Now that the hashing system is in place to
identify unique traces, even if some trace gets passed around the
different fuzzers, we should have no trouble filtering out the
duplicates.