rosa: add way to fix a seed for fuzzers
As observed during testing, when we reuse fuzzers (i.e. resume a fuzzing
campaign) there might be inconsistencies in the traces that are logged
by them. To fix this, we can use the same seed for all fuzzers in one
run; to still maintain the random aspect of fuzzing, we can generate
this seed randomly by default in rosa.