Refactor fuzzer launching logic
Instead of running two separate sets of fuzzers (one for the seed phase
and one of the detection phase), it's just much simpler to only run them
once, and to decide to arbitrarily draw the line between the two phases.
In reality, this distinction will be made thanks to coverage metrics
and/or a given amount of time (allocated to the seed phase).

This way, we elegantly avoid instrumentation variability between runs
and spurious memory/startup issues with the fuzzers.

The TUI is revamped to reflect these changes as well.