rosa: fix decisions for seed traces
In order to get more accurate metrics during the seed phase, the
decisions for the seed traces should be saved immediately instead of
waiting for the clustering phase. This has the disadvantage of not
knowing what cluster ID will correspond to what seed trace, but that's
not a deal breaker (we can look up the name of the trace in the
clusters/ directory). In this case, the cluster ID will be `"<none>"`,
indicating that the trace is considered to be a seed trace.