rosa: add option to not block on fuzzers
Sometimes the initialization process of the fuzzers takes way too long,
and we want to be able to have a quick seed phase. In order to make that
possible, add an option to ignore the fuzzer setup process and proceed
without waiting for all the fuzzers to be up and running.

This will make fuzzer startup errors less visible, in the interest of
having a shorter setup time.