Deduplicate findings based on discriminators
Rosa's findings are classified as suspicious based on the oracle
criterion. As such, it is possible that multiple input-trace pairs may
end up as findings (true/false positives), even though they have been
classified as findings for the same reasons.

For example, assuming we use `Criterion::SyscallsOnly` for the oracle
criterion, two input-trace pairs that exhibit the same behavior (i.e.,
have the same syscalls) may be classified as findings even though they
produce the exact same behavior.

We can leverage this information (called _discriminants_) to deduplicate
the findings. Thus, in the previous example, the two input-trace pairs
will still be saved, but in a manner that makes it clear that they
belong to the same behavior class.

Concretely, the `backdoors/` directory now contains other
sub-directories describing classes of discriminants based on the oracle
criterion. These sub-directories then contain each unique test input
that belongs to that class. When evaluating, the findings are
deduplicated in the same way, which means that in the previous example
we would only report one positive (instead of two).