Discriminant UID: incorporate cluster UID
Testing has showed that computing a discriminant UID and deduplicating
based on it results in some collisions in cases where there are too few
discriminants. In order to remedy this, the cluster UID is incorporated
into the discriminant UID, to ensure that we discriminate based on the
same cluster.

Naturally, this results in less deduplication (meaning that more false
positives may be reported), but it makes collisions between discriminant
UIDs less likely.