Move trace collection to fuzzer backend
With the oncoming modifications to be able to use ROSA with source
instrumentation fuzzing with AFL++, the backend needs to be able to
define how traces are collected.

Since the common scenario will involve "classic" trace dumps, we can
simply provide the existing behavior as a default implementation, and
let the backends potentially define alternate behaviors depending on
context.