Add support for standard AFL++ mode
To be able to use ROSA with standard AFL++ fuzzing using source-level
instrumentation, we need to be able to collect traces through an
alternate means, as we will not be able to generate trace dumps on the
AFL++ side in that case.

We can achieve this through `afl-showmap` and `strace` for edges and
syscalls respectively.