Optimize trace collection
Deduplicating new findings based on trace UID is necessary, but
suboptimal, as we have to pay the price of the computation of the UID
for every test input we come across.

In order to optimize this process, we will also keep track of the input
file paths seen so far. This way, even if something was discarded (i.e.,
not kept in the trace database), we will not spend time trying to
reevaluate it to arrive to the same conclusion and discard it again.
We use input *paths* instead of simple filenames to deal with the case
where we collect inputs from multiple fuzzer instances, with potentially
colliding names.

This seems to greatly improve the efficacy of trace collection,
especially when the computation of the UID is more expensive (e.g., in
the case of "standard" AFL++ fuzzing).