Pass config env vars to afl-showmap and strace
For consistency, environment variables declared in the configuration of
a fuzzer instance should also be passed when running the program through
afl-showmap and strace when collecting traces.

Note that we do not need to do this when running the program during
setup to get the coverage map size, as the program immediately prints
the size and quits (no need for, e.g., preloads).