rosa-filter-diff: add new (default) strategy
In order to make sure we do not miss more subtle backdoors (where the
difference is with the cluster, not the trace), make sure to take into
account *both* the trace and the cluster discriminants when taking a
new decision. This does not require a new inference; we simply look at
two things:

- Are some *trace discriminants* only present in the current trace
  (i.e., not in the base?)
- Are some *cluster discriminants* only absent in the current trace
  (i.e., not in the base?)

If both of these are true, then the corresponding input is suspicious.