AFL++ (source mode): do not bail on `afl-showmap` error
It has been observed that `afl-showmap` will spuriously crash with some
targets and test cases. These crashes might not be reproducible after
the fact (i.e., by running the same target and test case later), which
means that it's a shame that it brings `rosa` down with it.

In order to be more robust against this type of crash, we choose to
skip the trace (if allowed); if this is indeed a spurious crash, it will
get picked up later when `afl-showmap` is run again and doesn't crash.

This technically removes the protection that was added against
misconfigurations of `afl-showmap` (i.e., `AFL_MAP_SIZE` not being set
when it needs to be). However, in the case where a real problem exists
with the configuration, this will essentially result in all traces being
skipped, which still indirectly lets the user know there was a problem.