rosa-filter-diff: review backdoor criteria
We should not expect the **exact same** difference between the current
and base traces. Rather, if there is even a single system call that
persists (either relative to the trace or the cluster) in the current
trace, after the ones it has in common with the base trace have been
filtered out, then it is probably a backdoor.