Further cleanup in trace collection
In an effort to continue the cleanup/refactoring work from the previous
commits, each fuzzer instance is now assigned its own scratch directory,
in which it can store any temporary files.

Additionally, some functionality is removed from `trace.rs` to alleviate
it and separate it from the trace collection logic (which is part of the
backend).