Fix `rosa` not collecting all traces during the final pass
When all fuzzers have stopped, and if `--hard-stop` is not used, `rosa`
is supposed to pick up all remaining traces. However, this was not
previously done, as it would attempt to get one new trace and would
consider itself done if it already knew it, thus leaving some of the
available traces behind.

To solve this, a new method is introduced, which is aimed at handling
**all** remaining test inputs. This should not be an issue in terms of
performance as the fuzzers at this point are stopped.