AFL++ (standard mode): work around silly afl-showmap issue
`afl-showmap` errors out (with exit code 2) if the test input times out.
This is a legitimate way of doing things, except it also uses that exit
code for other reasons, which makes it unreliable for figuring out why
the exit code is non-zero.

In order to work around this while also failing for legitimate
`afl-showmap` errors, we will simply raise the timeout value to 5
seconds, which should be enough for slightly misbehaving inputs (PDF
files with Poppler in my tests). This is--of course--not bulletproof,
but I would rather not sacrifice error checking `afl-showmap`.