rosa.rs: ensure we collect >0 traces before clustering
Normally, we'd like to prioritize switching from seed collection to
clustering (and then backdoor detection) as quickly as possible. For
example, if the duration of phase one is X seconds, ideally at second
X+1 we are already running backdoor detection.

However, when we have the combination of a very slow target (which slows
down the fuzzers and also trace collection) and a very short phase one,
then we might switch over to clustering and backdoor detection with 0
traces. This is obviously problematic, as we need to have at least one
cluster to compare to.

To fix this, we ensure that at least one trace is present before
switching over to the clustering phase. This is not ideal as we won't
necessarily switch when promised, but `rosa.rs` can simply not work if
we don't have at least one cluster in place.