rosa-simulate: parallelize
Since a lot of the current experiments (specifically, the ablation
study ones) depend on `rosa-simulate`, and some targets are quite slow,
it would be very beneficial to parallelize `rosa-simulate`. We can do
this at least for the oracle inference part, as by that point the input
clusters have been formed, and all remaining traces can be evaluated
independently.