rosa: parallelize final trace collection
By default, `rosa` will collect traces for any and all remaining fuzzer
inputs at the end of the campaign. This was originally done
sequentially, in order to make sure to deduplicate traces (using a known
trace database). However, there is no need to do the collection itself
sequentially.

This commit introduces a change in the internal API for the AFL++
fuzzer backend, to allow for parallel trace collection. The API itself
internally deduplicates traces after the fact (which is fast), but the
collection itself is parallelized to avoid waiting too long at the end
of each campaign.