clustering.rs: fix bug with Criterion::EdgesAndSyscalls
Due to a bug during minimization of distance with the `EdgesAndSyscalls`
criterion, the syscall distance was overriding the edge distance.

This commit fixes the bug by keeping track of a "primary" and a
"secondary" distance. In the case of all criteria except
`EdgesAndSyscalls`, only the primary distance counts (i.e., edges in the
case of `EdgesOnly`, syscalls in the case of `SyscallsOnly`, the
smallest between edges and syscalls in the case of `EdgesOrSyscalls`).
In the case of `EdgesAndSyscalls`, we prioritize the smallest edge
distance, but if multiple traces *within a cluster* or *multiple
clusters* offer the same minimal edge distance, we use the syscall
distance as a tiebreaker. Hence, for this criterion, the edge distance
is the "primary" one and the syscall distance is the "secondary" one.